alert-hook/requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dataclasses==0.8
debtcollector==2.2.0
Flask==2.0.1
idna==2.10
importlib-metadata==4.6.1
importlib-resources==5.2.0
iso8601==0.1.14
itsdangerous==2.0.1
Jinja2==3.0.1
keystoneauth1==4.3.1
MarkupSafe==2.0.1
msgpack==1.0.2
netaddr==0.8.0
netifaces==0.11.0
os-service-types==1.7.0
oslo.config==8.7.0
oslo.i18n==5.0.1
oslo.serialization==4.1.0
oslo.utils==4.9.2
packaging==21.0
pbr==5.6.0
pyparsing==2.4.7
python-keystoneclient==4.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
stevedore==3.3.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
